requirements.txt
robin-stocks==1.5.4
pyyaml==5.3.1
numpy==1.19.4
pandas==1.2.0